+++ packages/gasket-resolve/CHANGELOG.md
@@ -1,6 +1,8 @@
 # `@gasket/resolve`
 
-- Validate that plugin objects are named
+### 5.3.1
+
+- Validate that plugin objects are named ([#156])
 
 ### 5.1.0
 
@@ -53,5 +55,6 @@
 [#105]: https://github.com/godaddy/gasket/pull/105
 [#141]: https://github.com/godaddy/gasket/pull/141
 [#142]: https://github.com/godaddy/gasket/pull/142
+[#156]: https://github.com/godaddy/gasket/pull/156
 
 [Loader]:/packages/gasket-resolve/docs/api.md#loader